% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httrex.R
\name{httrex}
\alias{httrex}
\title{Render a httrex}
\usage{
httrex(x = NULL, input = NULL, ...)
}
\arguments{
\item{x}{An expression. If not given, \code{httrex()} looks for code in
\code{input} or on the clipboard, in that order.}

\item{input}{Character. If has length one and lacks a terminating newline,
interpreted as the path to a file containing reprex code. Otherwise,
assumed to hold reprex code as character vector.}

\item{...}{Other options passed to \code{\link[reprex:reprex]{reprex::reprex()}}}
}
\description{
Render a httrex
}
