% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{hx_gadget}
\alias{hx_gadget}
\alias{hx_gadget_stop}
\title{Start/stop a httrex gadget}
\usage{
hx_gadget()

hx_gadget_stop()
}
\description{
A httrex gadget creates a shiny app that follows along your code and
tracks the httr calls that you make, providing details about your
interactions with a web API.
}
