% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbose.R
\name{hx_verbose}
\alias{hx_verbose}
\alias{with_hx_verbose}
\title{Give verbose output}
\usage{
hx_verbose(data_out = TRUE, data_in = FALSE, info = FALSE, ssl = FALSE)

with_hx_verbose(expr, ...)
}
\arguments{
\item{data_out}{Show data sent to server}

\item{data_in}{Show data received from server}

\item{info}{Show informational text from curl. This is mainly useful for debugging
https and auth problems, so is disabled by default}

\item{ssl}{Show even data sent/recieved over SSL connections?}

\item{expr}{An expression to be evaluated with verbose settings}

\item{...}{passed to \code{hx_verbose()}}
}
\value{
An \code{httr::config} object
}
\description{
A verbose connection provides much more information about the flow of
information between the client and server. \code{hx_verbose()} differs
from \code{\link[httr:verbose]{httr::verbose()}} because it prints during \code{httrex} / \code{reprex}
rendering.
}
\section{Overriding during \code{httrex} rendering}{

\code{httrex} allows for overriding options per knitr chunk using chunk options.
During a \code{httrex()} call the knitr options can be set by starting a line with
\verb{#+} and then setting the options to R values.

The available options are:
\itemize{
\item If the chunk label is "setup", no verbose output is provided. This allows
sensitive details to be kept confidential, such as a password when logging in.
\item \code{hx.data_in}, \code{hx.data_out}, \code{hx.info}, \code{hx.ssl} override the arguments to
\code{verbose()}.
}

Example code that could be sent to \code{httrex()}:\preformatted{#+ setup                # makes a chunk with label "setup" so there's no verbose output
library(crunch)
login()
#+                      # Nothing after `#+` returns to defaults
listDatasets()
#+ hx.data_out=TRUE     # Override so that `data_out` is printed
}
}

\section{Prefixes}{


\code{verbose()} uses the following prefixes to distinguish between
different components of the http messages:
\itemize{
\item \code{*} informative curl messages
\item \verb{->} headers sent (out)
\item \verb{>>} data sent (out)
\item \verb{*>} ssl data sent (out)
\item \verb{<-} headers received (in)
\item \verb{<<} data received (in)
\item \verb{<*} ssl data received (in)
}
}

